switch to uint32 internally from int

+++ internal/h7z/shift.go
@@ -1,5 +1,5 @@
 package h7z
 
-func urshift(number, bits int) int {
-	return int(uint(number) >> uint(bits))
+func urshift(number, bits uint32) uint32 {
+	return number >> bits
 }
